+++ src/common/global-locals.spec.ts
@@ -9,9 +9,13 @@ import { User } from '../users/user.entity';
 import { getPermissionsFromUser } from '../users/helpers/get-permissions-from-user.helper';
 import { UserPermission } from '../users/user-permission';
 import { getActingUser } from '../users/helpers/get-acting-user.helper';
+import { canChangeProfession } from '../users/helpers/can-change-profession';
+
+import professionFactory from '../testutils/factories/profession';
 
 jest.mock('../users/helpers/get-permissions-from-user.helper');
 jest.mock('../users/helpers/get-acting-user.helper');
+jest.mock('../users/helpers/can-change-profession');
 
 let request: DeepMocked<RequestWithAppSession>;
 let response: DeepMocked<Response>;
@@ -22,6 +26,7 @@ describe('globalLocals', () => {
     it('sets the expected variables', () => {
       createEnvironment();
       const user = createUserInEnvironment();
+      const profession = professionFactory.build();
 
       const permissions = [
         UserPermission.CreateOrganisation,
@@ -29,6 +34,7 @@ describe('globalLocals', () => {
       ];
 
       (getPermissionsFromUser as jest.Mock).mockReturnValue(permissions);
+      (canChangeProfession as jest.Mock).mockReturnValue(true);
 
       globalLocals(request, response, next);
 
@@ -36,8 +42,10 @@ describe('globalLocals', () => {
       expect(response.app.locals.user).toEqual(user);
       expect(response.app.locals.permissions).toEqual(permissions);
       expect(response.app.locals.currentUrl).toEqual('http://localhost/foo');
+      expect(response.app.locals.canChangeProfession(profession)).toEqual(true);
 
       expect(getPermissionsFromUser).toHaveBeenCalledWith(user);
+      expect(canChangeProfession).toHaveBeenCalledWith(user, profession);
     });
   });
 
@@ -51,8 +59,12 @@ describe('globalLocals', () => {
       expect(response.app.locals.user).toEqual(undefined);
       expect(response.app.locals.permissions).toEqual(undefined);
       expect(response.app.locals.currentUrl).toEqual('http://localhost/foo');
+      expect(
+        response.app.locals.canChangeProfession(professionFactory.build()),
+      ).toEqual(undefined);
 
       expect(getPermissionsFromUser).not.toHaveBeenCalled();
+      expect(canChangeProfession).not.toHaveBeenCalled();
     });
   });