+++ src/common/global-locals.ts
@@ -2,6 +2,8 @@ import { Response, NextFunction } from 'express';
 import { RequestWithAppSession } from './interfaces/request-with-app-session.interface';
 import { getPermissionsFromUser } from '../users/helpers/get-permissions-from-user.helper';
 import { getActingUser } from '../users/helpers/get-acting-user.helper';
+import { canChangeProfession } from '../users/helpers/can-change-profession';
+import { Profession } from '../professions/profession.entity';
 
 export function globalLocals(
   req: RequestWithAppSession,
@@ -16,6 +18,8 @@ export function globalLocals(
   res.app.locals.currentUrl = req.originalUrl;
   res.app.locals.infoMessages = req.flash('info');
   res.app.locals.successMessages = req.flash('success');
+  res.app.locals.canChangeProfession = (profession: Profession) =>
+    user && canChangeProfession(user, profession);
 
   next();
 }