+++ src/users/helpers/can-change-profession.spec.ts
@@ -1,15 +1,13 @@
-import { createDefaultMockRequest } from '../../testutils/factories/create-default-mock-request';
 import professionFactory from '../../testutils/factories/profession';
-import { checkCanViewProfession } from './check-can-view-profession';
+import { canChangeProfession } from './can-change-profession';
 import userFactory from '../../testutils/factories/user';
 import * as getActingUserModule from './get-acting-user.helper';
 import * as getTierOneOrganisationsFromProfessionModule from '../../professions/helpers/get-tier-one-organisations-from-profession.helper';
 import organisationFactory from '../../testutils/factories/organisation';
-import { UnauthorizedException } from '@nestjs/common';
 
-describe('checkCanViewProfession', () => {
+describe('canChangeProfession', () => {
   describe('when called as a service owner user', () => {
-    it('returns without throwing an Error', () => {
+    it('returns true', () => {
       const getActingUserSpy = jest.spyOn(getActingUserModule, 'getActingUser');
       const getTierOneOrganisationsFromProfessionSpy = jest.spyOn(
         getTierOneOrganisationsFromProfessionModule,
@@ -21,24 +19,15 @@ describe('checkCanViewProfession', () => {
 
       getActingUserSpy.mockReturnValue(user);
 
-      const request = createDefaultMockRequest({ user });
+      expect(canChangeProfession(user, profession)).toBeTruthy();
 
-      expect(() => {
-        checkCanViewProfession(request, profession);
-      }).not.toThrowError();
-
-      expect(getActingUserSpy).toBeCalledWith(request);
       expect(getTierOneOrganisationsFromProfessionSpy).not.toBeCalled();
     });
   });
 
   describe('when called as a non-service owner user', () => {
     describe('the profession has no tier-one organisations', () => {
-      it('throws an UnauthorizedException', () => {
-        const getActingUserSpy = jest.spyOn(
-          getActingUserModule,
-          'getActingUser',
-        );
+      it('returns false', () => {
         const getTierOneOrganisationsFromProfessionSpy = jest.spyOn(
           getTierOneOrganisationsFromProfessionModule,
           'getTierOneOrganisationsFromProfession',
@@ -50,16 +39,10 @@ describe('checkCanViewProfession', () => {
         });
         const profession = professionFactory.build();
 
-        getActingUserSpy.mockReturnValue(user);
         getTierOneOrganisationsFromProfessionSpy.mockReturnValue([]);
 
-        const request = createDefaultMockRequest({ user });
+        expect(canChangeProfession(user, profession)).toBeFalsy();
 
-        expect(() => {
-          checkCanViewProfession(request, profession);
-        }).toThrowError(UnauthorizedException);
-
-        expect(getActingUserSpy).toBeCalledWith(request);
         expect(getTierOneOrganisationsFromProfessionSpy).toBeCalledWith(
           profession,
         );
@@ -67,11 +50,7 @@ describe('checkCanViewProfession', () => {
     });
 
     describe("the profession has one tier-one organisation, which matches the user's", () => {
-      it('returns without throwing an Error', () => {
-        const getActingUserSpy = jest.spyOn(
-          getActingUserModule,
-          'getActingUser',
-        );
+      it('returns true', () => {
         const getTierOneOrganisationsFromProfessionSpy = jest.spyOn(
           getTierOneOrganisationsFromProfessionModule,
           'getTierOneOrganisationsFromProfession',
@@ -83,30 +62,20 @@ describe('checkCanViewProfession', () => {
         });
         const profession = professionFactory.build();
 
-        getActingUserSpy.mockReturnValue(user);
         getTierOneOrganisationsFromProfessionSpy.mockReturnValue([
           user.organisation,
         ]);
 
-        const request = createDefaultMockRequest({ user });
+        expect(canChangeProfession(user, profession)).toBeTruthy();
 
-        expect(() => {
-          checkCanViewProfession(request, profession);
-        }).not.toThrowError();
-
-        expect(getActingUserSpy).toBeCalledWith(request);
         expect(getTierOneOrganisationsFromProfessionSpy).toBeCalledWith(
           profession,
         );
       });
     });
 
-    describe("the profession has one tier-one organisation, which does not matches the user's", () => {
-      it('returns without throwing an Error', () => {
-        const getActingUserSpy = jest.spyOn(
-          getActingUserModule,
-          'getActingUser',
-        );
+    describe("the profession has one tier-one organisation, which does not match the user's", () => {
+      it('returns false', () => {
         const getTierOneOrganisationsFromProfessionSpy = jest.spyOn(
           getTierOneOrganisationsFromProfessionModule,
           'getTierOneOrganisationsFromProfession',
@@ -118,18 +87,12 @@ describe('checkCanViewProfession', () => {
         });
         const profession = professionFactory.build();
 
-        getActingUserSpy.mockReturnValue(user);
         getTierOneOrganisationsFromProfessionSpy.mockReturnValue([
           organisationFactory.build(),
         ]);
 
-        const request = createDefaultMockRequest({ user });
+        expect(canChangeProfession(user, profession)).toBeFalsy();
 
-        expect(() => {
-          checkCanViewProfession(request, profession);
-        }).toThrowError(UnauthorizedException);
-
-        expect(getActingUserSpy).toBeCalledWith(request);
         expect(getTierOneOrganisationsFromProfessionSpy).toBeCalledWith(
           profession,
         );
@@ -137,11 +100,7 @@ describe('checkCanViewProfession', () => {
     });
 
     describe("the profession has one tier-one organisation, which matches the user's, and another tier-one organisation, which does not", () => {
-      it('returns without throwing an Error', () => {
-        const getActingUserSpy = jest.spyOn(
-          getActingUserModule,
-          'getActingUser',
-        );
+      it('returns true', () => {
         const getTierOneOrganisationsFromProfessionSpy = jest.spyOn(
           getTierOneOrganisationsFromProfessionModule,
           'getTierOneOrganisationsFromProfession',
@@ -151,21 +110,16 @@ describe('checkCanViewProfession', () => {
           serviceOwner: false,
           organisation: organisationFactory.build(),
         });
+
         const profession = professionFactory.build();
 
-        getActingUserSpy.mockReturnValue(user);
         getTierOneOrganisationsFromProfessionSpy.mockReturnValue([
           user.organisation,
           organisationFactory.build(),
         ]);
 
-        const request = createDefaultMockRequest({ user });
+        expect(canChangeProfession(user, profession)).toBeTruthy();
 
-        expect(() => {
-          checkCanViewProfession(request, profession);
-        }).not.toThrowError();
-
-        expect(getActingUserSpy).toBeCalledWith(request);
         expect(getTierOneOrganisationsFromProfessionSpy).toBeCalledWith(
           profession,
         );